+++ src/cider/type/plan/Analyzer.h
@@ -40,6 +40,7 @@
 #include "cider/CiderException.h"
 #include "type/data/sqltypes.h"
 #include "type/plan/BinaryExpr.h"
+#include "type/plan/CaseExpr.h"
 #include "type/plan/ColumnExpr.h"
 #include "type/plan/ConstantExpr.h"
 #include "type/plan/DateExpr.h"
@@ -770,60 +771,6 @@ class AggExpr : public Expr {
   bool is_distinct;                     // true only if it is for COUNT(DISTINCT x)
   // APPROX_COUNT_DISTINCT error_rate, APPROX_QUANTILE quantile
   std::shared_ptr<Analyzer::Constant> arg1;
-};
-
-/*
- * @type CaseExpr
- * @brief the CASE-WHEN-THEN-ELSE expression
- */
-class CaseExpr : public Expr {
- public:
-  CaseExpr(const SQLTypeInfo& ti,
-           bool has_agg,
-           const std::list<std::pair<std::shared_ptr<Analyzer::Expr>,
-                                     std::shared_ptr<Analyzer::Expr>>>& w,
-           std::shared_ptr<Analyzer::Expr> e)
-      : Expr(ti, has_agg), expr_pair_list(w), else_expr(e) {}
-  const std::list<
-      std::pair<std::shared_ptr<Analyzer::Expr>, std::shared_ptr<Analyzer::Expr>>>&
-  get_expr_pair_list() const {
-    return expr_pair_list;
-  }
-  const Expr* get_else_expr() const { return else_expr.get(); }
-  const std::shared_ptr<Analyzer::Expr>& get_else_ref() const { return else_expr; }
-  std::shared_ptr<Analyzer::Expr> deep_copy() const override;
-  void check_group_by(
-      const std::list<std::shared_ptr<Analyzer::Expr>>& groupby) const override;
-  void group_predicates(std::list<const Expr*>& scan_predicates,
-                        std::list<const Expr*>& join_predicates,
-                        std::list<const Expr*>& const_predicates) const override;
-  void collect_rte_idx(std::set<int>& rte_idx_set) const override;
-  void collect_column_var(
-      std::set<const ColumnVar*, bool (*)(const ColumnVar*, const ColumnVar*)>&
-          colvar_set,
-      bool include_agg) const override;
-  std::shared_ptr<Analyzer::Expr> rewrite_with_targetlist(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override;
-  std::shared_ptr<Analyzer::Expr> rewrite_with_child_targetlist(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override;
-  std::shared_ptr<Analyzer::Expr> rewrite_agg_to_var(
-      const std::vector<std::shared_ptr<TargetEntry>>& tlist) const override;
-  bool operator==(const Expr& rhs) const override;
-  std::string toString() const override;
-  void find_expr(bool (*f)(const Expr*),
-                 std::list<const Expr*>& expr_list) const override;
-  std::shared_ptr<Analyzer::Expr> add_cast(const SQLTypeInfo& new_type_info) override;
-  void get_domain(DomainSet& domain_set) const override;
-
-  JITExprValue& codegen(CodegenContext& context) override;
-  ExprPtrRefVector get_children_reference() override;
-
- private:
-  std::list<std::pair<std::shared_ptr<Analyzer::Expr>, std::shared_ptr<Analyzer::Expr>>>
-      expr_pair_list;  // a pair of expressions for each WHEN expr1 THEN expr2.  expr1
-                       // must be of boolean type.  all expr2's must be of compatible
-                       // types and will be promoted to the common type.
-  std::shared_ptr<Analyzer::Expr> else_expr;  // expression for ELSE.  nullptr if omitted.
 };
 
 /*
